* Thu Feb 25 2010  Tim Hockin <thockin@google.com>   - Send %T (tsc), %C (cs), and %I (ip) on the socket and DBUS, for     completeness.   - Fix bug in client rules that caused an extra newline to be sent.   - Update mce_decode with new struct field names.
git-svn-id: https://mcedaemon.googlecode.com/svn/trunk@65 0d233194-844c-11dd-8f23-d9c8888c2974

+++ mced.h
@@ -110,7 +110,7 @@ struct mce {
 	uint32_t init_apic_id;	/* CPU initial APIC ID (-1UL for unknown) */
 	int32_t  socket;	/* CPU socket number (-1 for unknown) */
 	uint32_t mcg_cap;	/* MCG_CAP (0 for unknown) */
-	uint8_t  cs;		/* CPU code segment */
+	uint16_t cs;		/* CPU code segment */
 	uint8_t  bank;		/* MC bank */
 	int8_t   vendor;	/* CPU vendor (enum cpu_vendor) */
 } __attribute__ ((packed));
